refactor(data-management #780): extract components and composables to reduce file sizes
- Delete redundant instruction files (.copilot-instructions.md, copilot-instructions.md)
- Consolidate all rules in co2-calculator-rules.md.instructions.md
- Add RTK usage guidelines (SUPER IMPORTANT rule)
- Update component size limit from 300 to 500 lines
- Add pre-commit hook warning for components >500 lines
- Create GitHub issue template for component refactoring
Data Management Refactoring:
- DataEntryDialog: 624 → 36 lines (extracted to composable + molecule)
- ModuleConfig: 741 → 198 lines (extracted 4 composables + 4 molecules)
- SubmoduleConfig: 504 → 43 lines (extracted to composable + molecule)
New composables:
- useDataEntryDialog.ts (345 lines)
- useModuleConfig.ts (208 lines)
- useRecalculation.ts (205 lines)
- useSubmoduleConfig.ts (296 lines)
New molecules:
- DataEntryDialogContent.vue (270 lines)
- ModuleConfigSection.vue (101 lines)
- ModuleUploadsSection.vue (121 lines)
- ModuleRecalculationDialog.vue (86 lines)
- SubmoduleItem.vue (250 lines)
- ComputedFactorDialog.vue (57 lines)

Refs: #899
