--- .github/.copilot-instructions.md
@@ -1,76 +1,0 @@
-# Copilot Instructions for Frontend Senior Development
-
-## Overview
-
-This document provides guidelines and best practices for frontend development using TypeScript and Vue.js 3 (Composition API).
-
-## Tech Stack
-
-- Vue.js 3 (Composition API)
-- TypeScript 4.5+
-- Vite (build tool)
-- Pinia (state management)
-- Vue Router 4
-- Vitest/Jest (testing)
-- ESLint + Prettier
-
-## Project Structure
-
-```
-src/
-├── assets/         # Static assets
-├── components/     # Reusable Vue components
-│   ├── common/     # Shared components
-│   └── features/   # Feature-specific components
-├── composables/    # Reusable composition functions
-├── config/         # App configuration
-├── layouts/        # Page layouts
-├── pages/          # Page components
-├── router/         # Vue Router configuration
-├── services/       # API services
-├── stores/         # Pinia stores
-├── types/          # TypeScript types/interfaces
-└── utils/          # Utility functions
-```
-
-## TypeScript Best Practices
-
-- Define interfaces for all props, emits, and state
-- Use type inference where possible but explicitly type complex functions
-- Prefer interfaces for object definitions and type aliases for unions/intersections
-- Use Generics for reusable components and functions
-
-## Component Guidelines
-
-- Use Single File Components (SFC) with `<script setup lang="ts">`
-- Implement props validation with runtime and type checking
-- Keep components small and focused (max ~300 lines)
-- Extract reusable logic into composables
-
-## State Management
-
-- Use Pinia for global state
-- Prefer refs/reactive for local component state
-- Structure stores by feature domain
-- Implement proper typing for all store state, getters, and actions
-
-## Testing Strategy
-
-- Unit test composables and utilities
-- Component tests for complex components
-- Use test-driven development for critical features
-- Mock API calls and external dependencies
-
-## Performance Considerations
-
-- Implement code-splitting with dynamic imports
-- Use v-memo for list rendering optimization
-- Consider lazy loading for routes and large components
-- Properly memoize computed properties and event handlers
-
-## Common Patterns
-
-- Use provide/inject for deep prop passing
-- Implement proper error handling with try/catch
-- Create adapter layers for external API integration
-- Follow SOLID principles for maintainable code
